reorganizing sql.base folder
