+++ opsml/registry/sql/base/query_engine.py
@@ -18,8 +18,8 @@ from sqlalchemy.sql import FromClause, Select
 from sqlalchemy.sql.expression import ColumnElement
 
 from opsml.helpers.logging import ArtifactLogger
+from opsml.registry.sql.base.sql_schema import CardSQLTable, SQLTableGetter
 from opsml.registry.sql.semver import get_version_to_search
-from opsml.registry.sql.sql_schema import CardSQLTable, SQLTableGetter
 
 logger = ArtifactLogger.get_logger()
